rename directory cosmos_js to ethdbchain_js
