+++ src/ethdbchain_sig/constants.js
