+++ src/ethdbchain_sig/core.js
