+++ src/ethdbchain_sig/hash.js
